+++ src/HubSpot/HubSpot/HubSpotMarketplace.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Fungku\HubSpot\API;
+namespace HubSpot\HubSpot;
 
 class HubSpotMarketplace{
