+++ src/HubSpot/HubSpot/sdk.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Fungku\HubSpot\API;
+namespace HubSpot\HubSpot;
 
 /**
  * Copyright 2012 HubSpot, Inc.
@@ -21,9 +21,21 @@ namespace Fungku\HubSpot\API;
  */
 
 /**
- * An exception for HubSpot errors
- *
  * @author Christopher Hoult <chris.hoult@datasift.com>
  * @see https://github.com/chrishoult
  */
-class HubSpotException extends Exception {}
+
+/**
+ * Registers an autoloader for HubSpot_* classes
+ *
+ * @param string $className
+ */
+function hubspot_autoload($className) {
+	if (strpos($className, 'HubSpot_') == 0) {
+		$className = strtolower(str_replace('HubSpot_', '', $className));
+		$file = 'class.' . $className . '.php';
+		include __DIR__ . DIRECTORY_SEPARATOR . $file;
+	}
+}
+
+spl_autoload_register('hubspot_autoload');
